#267 [move] : behavior quest package 분리
