+++ app/src/main/java/com/byeboo/app/presentation/quest/behavior/navigation/QuestBehaviorNavigation.kt
@@ -7,8 +7,8 @@ import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.byeboo.app.core.model.quest.QuestType
 import com.byeboo.app.core.util.routeNavigation
-import com.byeboo.app.presentation.quest.behavior.QuestBehaviorCompleteRoute
-import com.byeboo.app.presentation.quest.behavior.QuestBehaviorWritingRoute
+import com.byeboo.app.presentation.quest.behavior.complete.QuestBehaviorCompleteRoute
+import com.byeboo.app.presentation.quest.behavior.writing.QuestBehaviorWritingRoute
 import com.byeboo.app.presentation.quest.behavior.navigation.QuestBehavior.QuestBehaviorComplete
 import com.byeboo.app.presentation.quest.behavior.navigation.QuestBehavior.QuestBehaviorWriting
